[llvm] Make reward incremental.
Background
----------

The goal of compiler optimization tasks is to minimize a cost function
`C(s)`. Examples of cost functions are runtime, code size, and power
consumption. For LLVM we have two cost functions: instruction count
and native code size.

To produce an incremental reward, we use the reduction in cost between
the current state `s_t` and a previous state `s_{t-1}`.

    R(s_t) = C(s_{t-1}) - C(s_t)

For many compiler tasks we have a default heuristic that we can
compare against. For example, for LLVM we can use the default pass
pipeline when run with optimizations (e.g. `-Oz`) as a baseline
policy. Running the baseline policy on the initial program state
`s_{t=0}` produces a baseline state `s_b` which can be used to produce
a baseline cost `C(s_b)`.

We use the reduction in cost achieved by this baseline policy to scale
the reward signal:

    R(s_t) = (C(s_{t-1}) - C(s_t)) / (C(s_{t=0}) - C(s_{b}))

Implementation
--------------

This patch makes all reward signals incremental. When a reward is
requested, the difference between the previous cost and current cost
are returned. To preserve the lazy evaluation of reward signals, the
previous cost is updated only when a reward is calculated, for
example:

    >>> env.reset()
    >>> r1 = env.reward["IrInstructionCount"]
    0
    >>> env.step(a1)
    >>> env.step(a2)
    >>> r2 = env.reward["IrInstructionCount"]
    10
    >>> r3 = env.reward["IrInstructionCount"]
    0

The `r1` value will be 0, since there is no change to an environment
where no actions have been made. The `r2` value will be the reduction
in instruction count achieved by the two steps `a1` and `a2`. The `r3`
value is 0, since no actions have been made between `r3` and the
previous reward computation `r2`.

Exposing absolute cost values
-----------------------------

Although reward signals are no incremental, users may still wish to
query the absolute values of cost functions. E.g. it may informative
to query the total number of instructions in the program, even if the
environment is not using instruction count as a reward.

This patch exposes each of the cost functions through the
ObservationSpace view. For example:

    >>> env.reset()
    >>> env.observation["IrInstructionCount"]
    496
    >>> env.observation["IrInstructionCountO3"]
    501

Here, the program currently contains 496 instructions. The program, if
it were compiled with -O3, would have contained 501 instructions.