Add a validation script.
This adds a script to validate agent results. The validation script
consumes a list of episodes and replays them, validating that the
correct reward is produced. There is also the option to validate that
a benchmark's semantics remains unchanged using a set of ad-hoc
validation callbacks, though these callbacks will be provided in a
later patch.

To validate agent results, produce a CSV in the format:

    benchmark,reward,commandline,walltime
    benchmark://cBench-v0/dijkstra,0.5,opt  input.bc -o output.bc,0.3
    benchmark://cBench-v0/crc32,0.5,opt  input.bc -o output.bc,0.3

where each row contains the outcome of an episode: the name of the
benchmark, the total reward for the episode, the sequence of actions
as a commandline, and the walltime. The `CompilerEnv.state.to_csv()`
method produces output in the required format. Then pass the CSV to
stdin of the validation script:

    $ python -m compiler_gym.bin.validate < results.csv \
        --env=llvm-v0 --reward=IrInstructionCount