Add examples/explore.py that explores the states that are reachable for a certain set of actions within a given episode length.
Partial sequences that end up in the same state are deduplicated,
dramatically reducing the number of states to be searched.

Duplicates are currently detected using a hashcode/fingerprint of the
IR, but it appears that there is some state in an env that is not
captured by printing the IR. Also, and perhaps related, the number of
duplicates identified is slightly different when using parallelism. So
there is some more digging to be done.

Conservatively set the changed flag to true in all LLVM envs, as the
env relies on it being accurate for correctness and it is not
reliable. Some passes say no change occurred when a change did occur.