Add an LlvmEnv.fork() method.
Calling env.fork() creates a duplicate environment with the current
state. For long episodes, this may be more efficient than resetting a
fresh state and replaying the episode.

This implementation is totally ad-hoc for LLVM. It creates a new
compiler service and makes a temporary benchmark out of the current
environment state by serializing to and reading from a bitcode file. A
lower overhead fork() could be implemented on the C++ service side by
creating a new LlvmEnvironment for the existing service, but that
would require the addition of a new API call. Let's see if fork()
performance becomes a key factor before pursuing that.