[llvm-ic-v0/cBench-v0] Leaderboard submission: Random Search.
This adds a pure random policy that selects actions randomly until a
fixed number of steps (the "patience" of the search) have been
evaluated without an improvement to reward. The search stops after a
predetermined amount of search time has elapsed.

The patience value was selected by running a random 200 searches on
programs from the `blas-v0` dataset and selecting the value that
produced the best average reward.