Revert "[ci] Persist the bazel cache."
The caching mechanism can cause more problems than it's
worth. Specifically, failures caused by races in cache writes and
corrupt bazel cache states. I would like to re-enable the CI caching
in the future but not until there is a UI button to manually delete
existing caches so that we can "start fresh".

This reverts commit d98fef1a3ba9b70e257a8cae4df564082c605b89.