[llvm] Remove LLVM binaries from wheel.
This patch removes the LLVM binaries from the shipped wheel. This is
to reduce the package size to be under the 100MB default maximum
imposed by PyPi.

Instead of shipping the files in the wheel, the LLVM binaries are
downloaded from an archive hosted by Facebook when needed. The
circumstances for needing them are: (1) starting an LLVM service, (2)
attempting to resolve the path to an LLVM binary.