Release v0.1.7.
This release introduces public leaderboards to track the performance
of user-submitted algorithms on compiler optimization tasks.

- Added a new `compiler_gym.leaderboard` package which contains
utilities for preparing leaderboard submissions.

- Added a LLVM instruction count leaderboard and seeded it with a
random search baseline.

- Added support for Python 3.9, extending the set of supported python
versions to 3.6, 3.7, 3.8, and 3.9.

- [llvm] Added a new `InstCount` observation space that contains the
counts of each type of instruction.

Build dependencies update notice
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

If you are building from source and upgrading from an older version of
CompilerGym, your build environment will need to be updated. The
easiest way to do that is to remove your existing conda environment
using `conda remove --name compiler_gym --all` and to repeat the steps
in "building from source" in the README.