[llvm] Strip module and source IDs in benchmarks.
This strips the ModuleID and source_filename values from the LLVM
module immediately after creation. Previously those values would leak
into the IR:

    >>> env = gym.make("llvm-v0")
    >>> env.reset(benchmark="cBench-v1/crc32")
    >>> print(env.ir)
    ; ModuleID = 'benchmark://cBench-v1/crc32'
    source_filename = "llvm-link"
    ...

This provided an opportunity for a learning algorithm to overfit to
the benchmarks by their name. Now, the ModuleID and source_filename
attributes are stripped, anonymizing the IRs:

    >>> print(env.ir)
    ; ModuleID = '-'
    source_filename = "-"
    ...