[datasets] Add Dataset __getitem__, __iter__, and __len__.
This adds python operator overloads that alias to existing methods to
make the Dataset class "feel" more like a regular python dictionary:

     >>> len(dataset)  # equivalent to dataset.n
     23

     >>> for benchmark in dataset:  # iterate over the class directly
     ...     pass

     >>> dataset["cbench-v1/crc32"]  # key a benchmark

This also renames Dataset.n to Dataset.size for consistent with other
containers like np.ndarray, and returns math.inf if the number of
benchmarks is infinite, not a negative integer. The advantage of
math.inf is that will poison any integer arithemtic, e.g.

     >>> sum(d.size for d in datasets)
     inf

if any one of the datasets has an infinite size. With a negative
number, this would instead compute a regular integer value.

Issue #45.