Implement CompilerEnv.seed()
The new Datasets API is more explicit about its source of randomness
when choosing benchmarks, so implement the standard gym env.seed()
method to allow users to set and record this seed.

One quick of the Env.seed method is that when called with no arguments
it  should choose a seed randomly and return it. Given that we do not
want to rely on any RNGs (that users may also have seeded), I chose
instead to use the walltime to generate a seed.

Issue #45.