[benchmarks] Add example C++/Python micro-benchmarks.
This extends the micro-benchmark script to record and report runtimes
for operations on the example gym service implementations in C++ and
Python. The idea is that this provides a useful reference when
evaluating the measurements of other environments. Since the example
services do no compilation, the benchmark performances are close to
a roofline.