[tests] Add a `make install-test-cov` target.
This uses pytest-cov to generate a test coverage report file
coverage.xml when `make install-test-cov` is run. This coverage report
is then uploaded to codecov.io.

The measured difference in runtime on the full test suite is
negligible, though I expect that is largely because the bottleneck if
not Python performance:

    $ time make install-test-cov
    ...
    make install-test-cov  6797.99s user 813.37s system 1767% cpu 7:10.70 total

    $ time make install-test
    ...
    make install-test  6752.54s user 826.92s system 1782% cpu 7:05.14 total

If we run only on the `tests/datasets` directory which comprise pure
python tests, the difference is more pronounced:

    make install-test-cov  19.25s user 10.46s system 1048% cpu 2.832 total
    make install-test  14.37s user 11.71s system 1415% cpu 1.843 total