Add a mechanism for sending arbitrary data payloads to the backend.
This patch adds a mechanism to the CompilerGym client/service
architecture so that the client can send arbitrary data messages to
the service, which can then generate a response. This can be used by
particular client/service implementations to provide a "back door" for
talking to the service in a way that doesn't conform to the normal RL
environment communication pattern.

github.com/facebookresearch/CompilerGym/issues/312