Compile and ship csmith as part of python package.
Currently, when the csmith dataset is first used, the source code is
downloaded, unpacked and compiled on the local machine. This can cause
issues for systems without a standard toolchain for building C code,
or for other systems with limitations on running arbitrary code, for
example, in Google's Colab notebook environments.

This patch adds csmith to the set of compiled C code used to build the
CompilerGym package. It adds the generated binaries and headers to the
pip wheel so that `pip install compiler_gym` will ship with a working
Csmith binary.

Fixes #345.