[www] Make the Flask API stateless.
Replace the start/end/undo/step endpoints with a single "step"
function that takes all of the variables needed to describe an
environment state (then benchmark, reward signal, and full actions
history).

This replaces the session-based API which is error prone and hard to
scale. Note that this new stateless API is only a proof-of-concept
implementation, as on every "step" it replays an entire episode. In
the future we will change this to maintain a pool of live environments
that can be used to serve stateless API requests more efficiently.