Check that environment is not `done` on observation request.
This patch improves the error reporting when computing an observation
fails. First, if the service produces an unexpected number of
observations, a ServiceError is raised, rather than the previous
assertion. Second, if the environment reports that it has reached a
terminal state, a ServiceError is raised, containing the error details
produced by the environment.