[util] Rewrite checkOutput() using boost::process.
The subprocess::Popen() library function was found to slow down when
called in a tight loop thousands of times. This patch reimplements the
compiler_gym::util::checkOutput() function to use boost::process
rather than subprocess::Popen.

On a test bench, the runtimes of this new implementation have been
found to be stable after ~50k runs.

Fixes #393.