[ci] Replace release test with a more simple workflow.
The benefit of the pip install test is mostly in catching install-time
issues, and not in running the code tests. This patch simplifies the
CI job to be much simpler and quicker: it simply pip installs the
package and checks that the main compiler_gym module can be imported.

This reduces the runtime of each workflow to 8x ~30sec jobs, so we can
use these savings to schedule a daily run.