[datasets] Add a weighted implementation of random_benchmark().
By default datasets are selected uniformly randomly in
Datasets.random_benchmark(). This means that datasets with a small
number of benchmarks will be overrepresented compared to datasets with
many benchmarks. To correct for this bias this patch adds a
weight_datasets_by_size argument which is equivalent to weighting the
dataset choice by the number of benchmarks in each dataset:

    >>> random.choices(datasets, weights=[len(p) for p in datasets])

Weighting the choice of datasets by their size means that datasets with
infinite sizes (such as random program generators) will be excluded from
sampling as their size is 0.

Fixes #422.