[util] Move create_logs_dir() to runfiles_path.
This moves the create_logs_dir() function into the runfile_path module
so that all of the functions for generating filesystem paths are in
the same module. This also improves the implementation by enabling
thread safe logging directory generation, and renames the function to
create_user_logs_dir() to better communicate the type of logs.