[llvm] Add API to construct a benchmark from clang invocation.
This adds a new API for constructing a benchmark from a command line
invocation of a compiler. For example:

    >>> benchmark = env.make_benchmark_from_command_line(
    ...     ["gcc", "in.c", "-DNDEBUG"]
    ... )

This provides control over the build, equivalent to:

    >>> benchmark = env.make_benchmark(["in.c"], copts=["-NDEBUG"])

The idea is to make it easier to integrate CompilerGym into an
existing build system, as it is normally possible to dump a verbose
log of all of the compiler invocations run during a build.