[core] Add a CompilerEnv.multistep() method.
CompilerEnv.step() currently accepts two types for the "action"
argument: a scalar action, or an iterable of actions. This kind of
type overloading does not work for list types.

This adds a new method, CompilerEnv.multistep(), that explicitly takes
takes an iterable sequence of actions. If you want to run multiple
actions in a single step, call this new method. Calling
CompilerEnv.step() with a list of actions still works, though with a
deprecation warning. In the v0.2.4 release support for lists of
actions in CompilerEnv.step() will be removed.

Fixes #610.