CMake: fix custom target dependencies
This fixes dependency issues where a dependency of a custom command
on a custom target guarantees only the order of building and does
not guarantee that the custom command would rerun every time
when the custom target is rebuilt.

The approach to fix this is for custom targets to record in the
OUTPUTS target property their output files. Then the dependent
custom command depends on the contents of the OUTPUTS property
using generator expressions. This creates a file level dependencies
that triggers rebuilds on change.