[tests] Minimize the fork regression tests.
This wrote a simple routine minimize_fork_regression_testcase() that
takes as input a forked regression test and iteratively attempts to
remove each of the actions that do not prevent the test from failing.

Running this on the five known regression tests simplifies them
dratically. In all but one case, they are simplified to a single
pre-fork and single post-fork action.

I'll dig more into the details of what's going on in later.

Issue #749.